optimazation failed... use t3a.xlarge *2
