+++ Kubernetes_study_folder/[EKS] 쿠버네티스 클러스터 생성/source_code/pipeline.py
@@ -1,126 +1,10 @@
 # -*- coding: utf-8 -*-
-# https://www.kubeflow.org/docs/components/pipelines/v1/sdk/connect-api/
 
-import re
-import requests
-from urllib.parse import urlsplit
-
-def get_istio_auth_session(url: str, username: str, password: str) -> dict:
-    """
-    Determine if the specified URL is secured by Dex and try to obtain a session cookie.
-    WARNING: only Dex `staticPasswords` and `LDAP` authentication are currently supported
-             (we default default to using `staticPasswords` if both are enabled)
-
-    :param url: Kubeflow server URL, including protocol
-    :param username: Dex `staticPasswords` or `LDAP` username
-    :param password: Dex `staticPasswords` or `LDAP` password
-    :return: auth session information
-    """
-    # define the default return object
-    auth_session = {
-        "endpoint_url": url,    # KF endpoint URL
-        "redirect_url": None,   # KF redirect URL, if applicable
-        "dex_login_url": None,  # Dex login URL (for POST of credentials)
-        "is_secured": None,     # True if KF endpoint is secured
-        "session_cookie": None  # Resulting session cookies in the form "key1=value1; key2=value2"
-    }
-
-    # use a persistent session (for cookies)
-    with requests.Session() as s:
-
-        ################
-        # Determine if Endpoint is Secured
-        ################
-        resp = s.get(url, allow_redirects=True)
-        if resp.status_code != 200:
-            raise RuntimeError(
-                f"HTTP status code '{resp.status_code}' for GET against: {url}"
-            )
-
-        auth_session["redirect_url"] = resp.url
-
-        # if we were NOT redirected, then the endpoint is UNSECURED
-        if len(resp.history) == 0:
-            auth_session["is_secured"] = False
-            return auth_session
-        else:
-            auth_session["is_secured"] = True
-
-        ################
-        # Get Dex Login URL
-        ################
-        redirect_url_obj = urlsplit(auth_session["redirect_url"])
-
-        # if we are at `/auth?=xxxx` path, we need to select an auth type
-        if re.search(r"/auth$", redirect_url_obj.path): 
-            
-            #######
-            # TIP: choose the default auth type by including ONE of the following
-            #######
-            
-            # OPTION 1: set "staticPasswords" as default auth type
-            redirect_url_obj = redirect_url_obj._replace(
-                path=re.sub(r"/auth$", "/auth/local", redirect_url_obj.path)
-            )
-            # OPTION 2: set "ldap" as default auth type 
-            # redirect_url_obj = redirect_url_obj._replace(
-            #     path=re.sub(r"/auth$", "/auth/ldap", redirect_url_obj.path)
-            # )
-            
-        # if we are at `/auth/xxxx/login` path, then no further action is needed (we can use it for login POST)
-        if re.search(r"/auth/.*/login$", redirect_url_obj.path):
-            auth_session["dex_login_url"] = redirect_url_obj.geturl()
-
-        # else, we need to be redirected to the actual login page
-        else:
-            # this GET should redirect us to the `/auth/xxxx/login` path
-            resp = s.get(redirect_url_obj.geturl(), allow_redirects=True)
-            if resp.status_code != 200:
-                raise RuntimeError(
-                    f"HTTP status code '{resp.status_code}' for GET against: {redirect_url_obj.geturl()}"
-                )
-
-            # set the login url
-            auth_session["dex_login_url"] = resp.url
-
-        ################
-        # Attempt Dex Login
-        ################
-        resp = s.post(
-            auth_session["dex_login_url"],
-            data={"login": username, "password": password},
-            allow_redirects=True
-        )
-        if len(resp.history) == 0:
-            raise RuntimeError(
-                f"Login credentials were probably invalid - "
-                f"No redirect after POST to: {auth_session['dex_login_url']}"
-            )
-
-        # store the session cookies in a "key1=value1; key2=value2" string
-        auth_session["session_cookie"] = "; ".join([f"{c.name}={c.value}" for c in s.cookies])
-
-    return auth_session
-    
-import kfp
-#세션 연결하기
-KUBEFLOW_ENDPOINT = "http://54.204.170.127:8080"
-KUBEFLOW_USERNAME = "user@example.com"
-KUBEFLOW_PASSWORD = "12341234"
-
-auth_session = get_istio_auth_session(
-    url=KUBEFLOW_ENDPOINT,
-    username=KUBEFLOW_USERNAME,
-    password=KUBEFLOW_PASSWORD
-)
-
-client = kfp.Client(host=f"{KUBEFLOW_ENDPOINT}/pipeline", cookies=auth_session["session_cookie"])
-print(client.list_experiments())
-"""
-"""
 import kfp
 
 import kfp.dsl as dsl
+
+
 '''
 def simple_echo(i : int) -> int:
     return i
@@ -219,5 +103,27 @@ def calc_pipeline(
 
 
 arguments = {'a':'7'}#, 'b':'8'}
+
+
+
+
+
+from istio_auth import get_istio_auth_session
+
+#세션 연결하기
+KUBEFLOW_ENDPOINT = "http://54.204.170.127:8080"
+KUBEFLOW_USERNAME = "user@example.com"
+KUBEFLOW_PASSWORD = "12341234"
+
+auth_session = get_istio_auth_session(
+    url=KUBEFLOW_ENDPOINT,
+    username=KUBEFLOW_USERNAME,
+    password=KUBEFLOW_PASSWORD
+)
+
+client = kfp.Client(host=f"{KUBEFLOW_ENDPOINT}/pipeline", cookies=auth_session["session_cookie"])
+print(client.list_experiments())
+
+
 
 client.create_run_from_pipeline_func(calc_pipeline, arguments=arguments)
